+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.30.3
 ### Bugfix
 * fix prompt that return string or promise (in animation and set_prompt) [#724](https://github.com/jcubic/jquery.terminal/issues/724)
+* scroll to bottom for each line in multiline typing animation
 
 ## 2.30.2
 ### Bugfix